+++ NRG-APP/src/main/java/com/nrg/services/VehicleService.java
@@ -2,12 +2,10 @@ package com.nrg.services;
 
 import java.util.List;
 
-import com.nrg.models.User;
-import com.nrg.models.Vechicle;
 import com.nrg.models.VehicleDetails;
 
 public interface VehicleService {
 
-	List<VehicleDetails> getVehiclesOfUser(Long userId);
+	List<VehicleDetails> getVehiclesOfUser(Integer userId);
 	
 }
